Initial refactor for bnotest and pwmtest
