+++ nyansat/host/shell/command_invoker.py
@@ -1,5 +1,7 @@
 import ast
+from dataclasses import dataclass
 import json
+from typing import List
 
 from nyansat.host.shell.nyan_pyboard import NyanPyboard
 from nyansat.host.shell.errors import *
@@ -208,3 +210,141 @@ class CommandInvoker(NyanPyboard):
             return ret
         except PyboardError as e:
             raise NotRespondingError(str(e))
+
+    def bno_diagnostics(self, sda, scl):
+        """
+        Create a BNO controller object for the given I2C sda/scl configuration. Uses the default
+        value of 40 for the BNO055 I2C address.
+        :param sda: Pin number for sda
+        :param scl: Pin number for scl
+        :return: A BnoTestDiagnostics object containing relevant T/F information about the setup
+        """
+        i2c_bus_scannable = False
+        i2c_addresses = []
+        bno_object_created = False
+        bno_object_calibrated = False
+
+        # Test scanning I2C bus
+        try:
+            addresses = self.i2c_scan(sda, scl)
+        except PyboardError:
+            return BnoTestDiagnostics(
+                i2c_bus_scannable,
+                i2c_addresses,
+                bno_object_created,
+                bno_object_calibrated
+            )
+        i2c_bus_scannable = True
+
+        # Test what's on the I2C bus and their addresses
+        try:
+            i2c_addresses = [int(n) for n in addresses.strip('] [').split(', ')]
+            if not i2c_addresses:
+                return BnoTestDiagnostics(
+                    i2c_bus_scannable,
+                    i2c_addresses,
+                    bno_object_created,
+                    bno_object_calibrated
+                )
+        except ValueError:
+            return BnoTestDiagnostics(
+                i2c_bus_scannable,
+                i2c_addresses,
+                bno_object_created,
+                bno_object_calibrated
+            )
+
+        # Test creating BNO object
+        try:
+            self.exec_("from imu.imu_bno055 import Bno055ImuController")
+            self.exec_("bno = Bno055ImuController(i2c)")
+        except PyboardError:
+            return BnoTestDiagnostics(
+                i2c_bus_scannable,
+                i2c_addresses,
+                bno_object_created,
+                bno_object_calibrated
+            )
+        bno_object_created = True
+
+        # Test calibration status of BNO object
+        try:
+            calibration_status = json.loads(self.eval_string_expr("bno.get_calibration_status()"))
+            bno_object_calibrated = calibration_status['system'] > 0
+        except PyboardError:
+            bno_object_calibrated = False
+
+        return BnoTestDiagnostics(
+            i2c_bus_scannable,
+            i2c_addresses,
+            bno_object_created,
+            bno_object_calibrated
+        )
+
+    def pwm_diagnostics(self, sda, scl):
+        """
+        Create a PCA9685 controller object for the given I2C sda/scl configuration. Uses the default
+        value of 40 for the controller's address.
+        :param sda: Pin number for sda
+        :param scl: Pin number for scl
+        :return: A PwmTestDiagnostics object containing relevant T/F information about the setup
+        """
+        i2c_bus_scannable = False
+        i2c_addresses = []
+        pca_object_created = False
+
+        # Test scanning I2C bus
+        try:
+            addresses = self.i2c_scan(sda, scl)
+        except PyboardError:
+            return PwmTestDiagnostics(
+                i2c_bus_scannable,
+                i2c_addresses,
+                pca_object_created
+            )
+        i2c_bus_scannable = True
+
+        # Test what's on the I2C bus and their addresses
+        try:
+            i2c_addresses = [int(n) for n in addresses.strip('] [').split(', ')]
+            if not i2c_addresses:
+                return PwmTestDiagnostics(
+                    i2c_bus_scannable,
+                    i2c_addresses,
+                    pca_object_created
+                )
+        except ValueError:
+            return PwmTestDiagnostics(
+                i2c_bus_scannable,
+                i2c_addresses,
+                pca_object_created
+            )
+
+        # Test creating BNO object
+        try:
+            self.exec_("from motor.motor_pca9685 import Pca9685Controller")
+            self.exec_("pca = Pca9685Controller(i2c)")
+            pca_object_created = True
+        except PyboardError:
+            pca_object_created = False
+
+        return PwmTestDiagnostics(
+            i2c_bus_scannable,
+            i2c_addresses,
+            pca_object_created
+        )
+
+@dataclass
+class BnoTestDiagnostics:
+    """Store diagnostic T/F values for BNO test. Used in the handling for 'bnotest' command."""
+    i2c_bus_scannable: bool
+    i2c_addresses: List[int]
+    bno_object_created: bool
+    bno_object_calibrated: bool
+
+@dataclass
+class PwmTestDiagnostics:
+    """Store diagnostic T/F values for PWM test. Used in the handling for 'pwmtest' command."""
+    i2c_bus_scannable: bool
+    i2c_addresses: List[int]
+    pca_object_created: bool
